Incorporating feedback.
* Set tsconfig.json#rootDir to `.`
* Created files to house common test functionality
* Simplified unit test looping
